rename ty module to signature
This can contain the "parsed form" of function signatures etc.
